Update ripple effect & usage

+++ PixelController/lib/patternGeneration/utils.cpp
@@ -14,12 +14,12 @@ unsigned char *bufferPtr = &bufferPattern[0][0][0];
 /**
  * Calculates the distance between two (x, y) points provided.
  */
-uint8_t calculateDistance(uint8_t center_x, uint8_t center_y, uint8_t x, uint8_t y)
+uint16_t calculateDistance(uint16_t center_x, uint16_t center_y, uint16_t x, uint16_t y)
 {
     // Calculate Euclidean distance from center point to point (x, y)
     int dx = x - center_x;
     int dy = y - center_y;
-    return static_cast<uint8_t>(sqrt(dx * dx + dy * dy));
+    return static_cast<uint16_t>(sqrt(dx * dx + dy * dy));
 }
 
 /**
